Move core managers into managers package
