+++ mainappsrc/managers/diagram_clipboard_manager.py
