+++ mainappsrc/managers/project_properties_manager.py
