+++ mainappsrc/version.py
@@ -18,6 +18,6 @@
 
 """Project version information."""
 
-VERSION = "0.2.105"
+VERSION = "0.2.106"
 
 __all__ = ["VERSION"]
