+++ tests/test_app_undo_move_coalesce.py
@@ -20,7 +20,7 @@ import unittest
 import json
 
 from AutoML import AutoMLApp
-from mainappsrc.core.undo_manager import UndoRedoManager
+from mainappsrc.managers.undo_manager import UndoRedoManager
 
 
 class AppUndoMoveCoalesceTests(unittest.TestCase):
