+++ tests/test_cross_diagram_clipboard.py
@@ -26,7 +26,7 @@ import types
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from AutoML import AutoMLApp
-from mainappsrc.core.diagram_clipboard_manager import DiagramClipboardManager
+from mainappsrc.managers.diagram_clipboard_manager import DiagramClipboardManager
 from gui.architecture import SysMLDiagramWindow, _get_next_id, SysMLObject, ARCH_WINDOWS
 from gui.gsn_diagram_window import GSNDiagramWindow, GSN_WINDOWS
 from mainappsrc.models.gsn import GSNNode, GSNDiagram
