+++ tests/test_governance_undo.py
@@ -33,7 +33,7 @@ sys.modules.setdefault("PIL.ImageFont", types.ModuleType("PIL.ImageFont"))
 from AutoML import AutoMLApp
 from gui.architecture import GovernanceDiagramWindow, SysMLObject
 from mainappsrc.models.sysml.sysml_repository import SysMLRepository
-from mainappsrc.core.undo_manager import UndoRedoManager
+from mainappsrc.managers.undo_manager import UndoRedoManager
 
 
 def test_governance_diagram_undo_redo_work_product():
