+++ tests/test_gsn_clone_independence.py
@@ -30,7 +30,7 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from mainappsrc.models.gsn import GSNNode, GSNDiagram
 from AutoML import AutoMLApp
-from mainappsrc.core.diagram_clipboard_manager import DiagramClipboardManager
+from mainappsrc.managers.diagram_clipboard_manager import DiagramClipboardManager
 
 
 def _make_app(root, diagram):
