+++ tests/test_gsn_clone_reject.py
@@ -25,7 +25,7 @@ sys.path.append(str(Path(__file__).resolve().parents[1]))
 from mainappsrc.models.gsn.nodes import GSNNode
 from mainappsrc.models.gsn.diagram import GSNDiagram
 from AutoML import AutoMLApp, AutoML_Helper
-from mainappsrc.core.diagram_clipboard_manager import DiagramClipboardManager
+from mainappsrc.managers.diagram_clipboard_manager import DiagramClipboardManager
 from gui.controls import messagebox
