+++ tests/test_gsn_window_focus.py
@@ -20,7 +20,7 @@ import types
 import weakref
 
 from AutoML import AutoMLApp
-from mainappsrc.core.diagram_clipboard_manager import DiagramClipboardManager
+from mainappsrc.managers.diagram_clipboard_manager import DiagramClipboardManager
 from gui.gsn_diagram_window import GSNDiagramWindow, GSNNode, GSNDiagram, GSN_WINDOWS
