+++ tests/test_no_empty_fta_tab.py
@@ -31,7 +31,7 @@ sys.modules.setdefault("PIL.ImageTk", types.ModuleType("PIL.ImageTk"))
 
 from AutoML import AutoMLApp
 import AutoML
-from mainappsrc.core.undo_manager import UndoRedoManager
+from mainappsrc.managers.undo_manager import UndoRedoManager
 from mainappsrc.managers.project_manager import ProjectManager
 
 class DummyNotebook:
